fix(warnings): 清零 Thunder 自身编译告警 (#7)
修复项:
- FileLogger.cpp: fprintf(fp, str.c_str()) → fputs(fp, str) 消除 -Wformat-security
- StringCoder.cpp: y 变量初始化为 0，消除 -Wmaybe-uninitialized
- Manager.cpp / Worker.cpp: int iPosXxx → size_t 消除 -Woverflow(npos→int)
- StepTellWorker.cpp / DbOperator.cpp: 格式符与类型匹配，消除 -Wformat=
- Labor.cpp: strerror_r 返回值赋给 errStr(void)，消除 -Wunused-result
- CommonUtils.hpp: 移除 U+202C 双向控制符，消除 -Wbidi-chars
- base64.c / HttpUrlCoder.cpp: 移除 register 关键字，消除 -Wregister
- MysqlDbi.cpp: strncpy→snprintf 消除 -Wstringop-truncation
- FileUtil.cpp: strncpy+strncat → snprintf 合并，消除 -Wstringop-truncation + -Waddress
- IpUtil.cpp: 移除未使用的 ptr 变量，消除 -Wunused-variable
- CurlClient.{hpp,cpp}: 已弃用 CURLOPT 加 #pragma suppress，消除 -Wdeprecated-declarations
- test_codec_http.cpp: params_size() static_cast<size_t>，消除 -Wnarrowing
- test_connector_tcp.cpp: ev_loop_new 加 pragma，消除 -Wdeprecated-enum-enum-conversion

结果: Thunder 自身代码 0 warning, 0 error
288/288 ctest + 60/60 pytest + E2E 全通过

+++ code/Net/src/labor/Manager.cpp
@@ -940,13 +940,13 @@ bool Manager::SetConnectIdentify(const tagMsgShell& stMsgShell, const std::strin
 
 bool Manager::ParseAutoSendTarget(const std::string& strIdentify, AutoSendTarget& target)
 {
-    int iPosIpPortSeparator = strIdentify.find(':');
+    size_t iPosIpPortSeparator = strIdentify.find(':');
     if (iPosIpPortSeparator == std::string::npos)
     {
         LOG4_ERROR("iPosIpPortSeparator == std::string::npos");
         return false;
     }
-    int iPosPortWorkerIndexSeparator = strIdentify.rfind('.');
+    size_t iPosPortWorkerIndexSeparator = strIdentify.rfind('.');
     // 当 identify 不含 .worker_index 后缀时，默认 worker_index=0
     if (iPosPortWorkerIndexSeparator == std::string::npos
         || iPosPortWorkerIndexSeparator < iPosIpPortSeparator)
